Add tab marker lines to Text Editor. Ongoing collapse/expand blocks.
