+++ src/robotide/editor/customsourceeditor.py
@@ -16,44 +16,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from __future__ import absolute_import
 
-#
-# from StyledTextCtrl_2 import PythonSTC
-# -----------------------------------------------------------------
-# Define Python style
-import keyword
 import os
 import sys
 
 import wx
 import wx.stc as stc
+from .pythoneditor import PythonSTC
 from wx import Colour
-
-if wx.Platform == '__WXMSW__':
-    faces = {'times': 'Times New Roman',
-             'mono': 'Courier New',
-             'helv': 'Arial',
-             'other': 'Comic Sans MS',
-             'size': 10,
-             'size2': 8,
-             }
-elif wx.Platform == '__WXMAC__':
-    faces = {'times': 'Times New Roman',
-             'mono': 'Monaco',
-             'helv': 'Arial',
-             'other': 'Comic Sans MS',
-             'size': 12,
-             'size2': 10,
-             }
-else:
-    faces = {'times': 'Times',
-             'mono': 'Courier',
-             'helv': 'Helvetica',
-             'other': 'new century schoolbook',
-             'size': 12,
-             'size2': 10,
-             }
 
 # ---------------------------------------------------------------------------
 # This is how you pre-establish a file filter so that the dialog
@@ -65,292 +35,6 @@ wildcard = "All files (*.*)|*.*|"                \
            "Robot Framework (*.txt)|*.txt|" \
            "YAML file (*.yaml)|*.yaml"
 # ----------------------------------------------------------------------
-
-
-class PythonSTC(stc.StyledTextCtrl):
-
-    fold_symbols = 2
-
-    def __init__(self, parent, idd, pos=wx.DefaultPosition, size=wx.DefaultSize, style=0):
-        stc.StyledTextCtrl.__init__(self, parent, idd, pos, size, style)
-
-        self.CmdKeyAssign(ord('B'), stc.STC_SCMOD_CTRL, stc.STC_CMD_ZOOMIN)
-        self.CmdKeyAssign(ord('N'), stc.STC_SCMOD_CTRL, stc.STC_CMD_ZOOMOUT)
-
-        self.SetLexer(stc.STC_LEX_PYTHON)
-        self.SetKeyWords(0, " ".join(keyword.kwlist))
-
-        self.SetProperty("fold", "1")
-        self.SetProperty("tab.timmy.whinge.level", "1")
-        self.SetMargins(0, 0)
-
-        self.SetViewWhiteSpace(False)
-        self.SetEdgeMode(stc.STC_EDGE_BACKGROUND)
-        self.SetEdgeColumn(78)
-
-        # Set up a margin to hold fold markers
-        # self.SetFoldFlags(16)  ###  WHAT IS THIS VALUE?  WHAT ARE THE OTHER FLAGS?  DOES IT MATTER?
-        self.SetMarginType(2, stc.STC_MARGIN_SYMBOL)
-        self.SetMarginMask(2, stc.STC_MASK_FOLDERS)
-        self.SetMarginSensitive(2, True)
-        self.SetMarginWidth(2, 12)
-
-        if self.fold_symbols == 0:
-            # Arrow pointing right for contracted folders, arrow pointing down for expanded
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDEROPEN,    stc.STC_MARK_ARROWDOWN, "black", "black")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDER,        stc.STC_MARK_ARROW, "black", "black")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDERSUB,     stc.STC_MARK_EMPTY, "black", "black")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDERTAIL,    stc.STC_MARK_EMPTY, "black", "black")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDEREND,     stc.STC_MARK_EMPTY,     "white", "black")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDEROPENMID, stc.STC_MARK_EMPTY,     "white", "black")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDERMIDTAIL, stc.STC_MARK_EMPTY,     "white", "black")
-
-        elif self.fold_symbols == 1:
-            # Plus for contracted folders, minus for expanded
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDEROPEN,    stc.STC_MARK_MINUS, "white", "black")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDER,        stc.STC_MARK_PLUS,  "white", "black")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDERSUB,     stc.STC_MARK_EMPTY, "white", "black")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDERTAIL,    stc.STC_MARK_EMPTY, "white", "black")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDEREND,     stc.STC_MARK_EMPTY, "white", "black")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDEROPENMID, stc.STC_MARK_EMPTY, "white", "black")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDERMIDTAIL, stc.STC_MARK_EMPTY, "white", "black")
-
-        elif self.fold_symbols == 2:
-            # Like a flattened tree control using circular headers and curved joins
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDEROPEN,    stc.STC_MARK_CIRCLEMINUS,          "white", "#404040")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDER,        stc.STC_MARK_CIRCLEPLUS,           "white", "#404040")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDERSUB,     stc.STC_MARK_VLINE,                "white", "#404040")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDERTAIL,    stc.STC_MARK_LCORNERCURVE,         "white", "#404040")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDEREND,     stc.STC_MARK_CIRCLEPLUSCONNECTED,  "white", "#404040")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDEROPENMID, stc.STC_MARK_CIRCLEMINUSCONNECTED, "white", "#404040")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDERMIDTAIL, stc.STC_MARK_TCORNERCURVE,         "white", "#404040")
-
-        elif self.fold_symbols == 3:
-            # Like a flattened tree control using square headers
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDEROPEN,    stc.STC_MARK_BOXMINUS,          "white", "#808080")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDER,        stc.STC_MARK_BOXPLUS,           "white", "#808080")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDERSUB,     stc.STC_MARK_VLINE,             "white", "#808080")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDERTAIL,    stc.STC_MARK_LCORNER,           "white", "#808080")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDEREND,     stc.STC_MARK_BOXPLUSCONNECTED,  "white", "#808080")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDEROPENMID, stc.STC_MARK_BOXMINUSCONNECTED, "white", "#808080")
-            self.MarkerDefine(stc.STC_MARKNUM_FOLDERMIDTAIL, stc.STC_MARK_TCORNER,           "white", "#808080")
-
-        self.Bind(stc.EVT_STC_UPDATEUI, self.on_update_ui)
-        self.Bind(stc.EVT_STC_MARGINCLICK, self.on_margin_click)
-        self.Bind(wx.EVT_KEY_DOWN, self.on_key_pressed)
-
-        # Make some styles,  The lexer defines what each style is used for, we
-        # just have to define what each style looks like.  This set is adapted from
-        # Scintilla sample property files.
-
-        # Global default styles for all languages
-        self.StyleSetSpec(stc.STC_STYLE_DEFAULT,     "face:%(helv)s,size:%(size)d" % faces)
-        self.StyleClearAll()  # Reset all to be like the default
-
-        # Global default styles for all languages
-        self.StyleSetSpec(stc.STC_STYLE_DEFAULT,     "face:%(helv)s,size:%(size)d" % faces)
-        self.StyleSetSpec(stc.STC_STYLE_LINENUMBER,  "back:#C0C0C0,face:%(helv)s,size:%(size2)d" % faces)
-        self.StyleSetSpec(stc.STC_STYLE_CONTROLCHAR, "face:%(other)s" % faces)
-        self.StyleSetSpec(stc.STC_STYLE_BRACELIGHT,  "fore:#FFFFFF,back:#0000FF,bold")
-        self.StyleSetSpec(stc.STC_STYLE_BRACEBAD,    "fore:#000000,back:#FF0000,bold")
-
-        # Python styles
-        # Default
-        self.StyleSetSpec(stc.STC_P_DEFAULT, "fore:#000000,face:%(helv)s,size:%(size)d" % faces)
-        # Comments
-        self.StyleSetSpec(stc.STC_P_COMMENTLINE, "fore:#007F00,face:%(other)s,size:%(size)d" % faces)
-        # Number
-        self.StyleSetSpec(stc.STC_P_NUMBER, "fore:#007F7F,size:%(size)d" % faces)
-        # String
-        self.StyleSetSpec(stc.STC_P_STRING, "fore:#7F007F,face:%(helv)s,size:%(size)d" % faces)
-        # Single quoted string
-        self.StyleSetSpec(stc.STC_P_CHARACTER, "fore:#7F007F,face:%(helv)s,size:%(size)d" % faces)
-        # Keyword
-        self.StyleSetSpec(stc.STC_P_WORD, "fore:#00007F,bold,size:%(size)d" % faces)
-        # Triple quotes
-        self.StyleSetSpec(stc.STC_P_TRIPLE, "fore:#7F0000,size:%(size)d" % faces)
-        # Triple double quotes
-        self.StyleSetSpec(stc.STC_P_TRIPLEDOUBLE, "fore:#7F0000,size:%(size)d" % faces)
-        # Class name definition
-        self.StyleSetSpec(stc.STC_P_CLASSNAME, "fore:#0000FF,bold,underline,size:%(size)d" % faces)
-        # Function or method name definition
-        self.StyleSetSpec(stc.STC_P_DEFNAME, "fore:#007F7F,bold,size:%(size)d" % faces)
-        # Operators
-        self.StyleSetSpec(stc.STC_P_OPERATOR, "bold,size:%(size)d" % faces)
-        # Identifiers
-        self.StyleSetSpec(stc.STC_P_IDENTIFIER, "fore:#000000,face:%(helv)s,size:%(size)d" % faces)
-        # Comment-blocks
-        self.StyleSetSpec(stc.STC_P_COMMENTBLOCK, "fore:#7F7F7F,size:%(size)d" % faces)
-        # End of line where string is not closed
-        self.StyleSetSpec(stc.STC_P_STRINGEOL, "fore:#000000,face:%(mono)s,back:#E0C0E0,eol,size:%(size)d" % faces)
-
-        self.SetCaretForeground("BLUE")
-
-        # register some images for use in the AutoComplete box.
-        # self.RegisterImage(1, Smiles.GetBitmap())  # DEBUG was images.
-        self.RegisterImage(1, wx.ArtProvider.GetBitmap(wx.ART_FLOPPY, size=(16, 16)))
-        self.RegisterImage(2, wx.ArtProvider.GetBitmap(wx.ART_NEW, size=(16, 16)))
-        self.RegisterImage(3, wx.ArtProvider.GetBitmap(wx.ART_COPY, size=(16, 16)))
-
-    def on_key_pressed(self, event):
-        if self.CallTipActive():
-            self.CallTipCancel()
-        key = event.GetKeyCode()
-
-        if key == 32 and event.ControlDown():
-            pos = self.GetCurrentPos()
-
-            # Tips
-            if event.ShiftDown():
-                self.CallTipSetBackground("yellow")
-                self.CallTipShow(pos, 'lots of of text: blah, blah, blah\n\n'
-                                 'show some suff, maybe parameters..\n\n'
-                                 'fubar(param1, param2)')
-            # Code completion
-            else:
-                kw = list(keyword.kwlist[:])
-                kw.append("zzzzzz?2")
-                kw.append("aaaaa?2")
-                kw.append("__init__?3")
-                kw.append("zzaaaaa?2")
-                kw.append("zzbaaaa?2")
-                kw.append("this_is_a_longer_value")
-                # kw.append("this_is_a_much_much_much_much_much_much_much_longer_value")
-
-                kw.sort()  # Python sorts are case-sensitive
-                self.AutoCompSetIgnoreCase(False)  # so this needs to match
-
-                # Images are specified with an appended "?type"
-                for i in range(len(kw)):
-                    if kw[i] in keyword.kwlist:
-                        kw[i] = kw[i] + "?1"
-
-                self.AutoCompShow(0, " ".join(kw))
-        else:
-            event.Skip()
-
-    def on_update_ui(self, evt):
-        _ = evt
-        # check for matching braces
-        brace_at_caret = -1
-        brace_opposite = -1
-        char_before = style_before = None
-        caret_pos = self.GetCurrentPos()
-
-        if caret_pos > 0:
-            char_before = self.GetCharAt(caret_pos - 1)
-            style_before = self.GetStyleAt(caret_pos - 1)
-
-        # check before
-        if char_before and chr(char_before) in "[]{}()" and style_before == stc.STC_P_OPERATOR:
-            brace_at_caret = caret_pos - 1
-
-        # check after
-        if brace_at_caret < 0:
-            char_after = self.GetCharAt(caret_pos)
-            style_after = self.GetStyleAt(caret_pos)
-
-            if char_after and chr(char_after) in "[]{}()" and style_after == stc.STC_P_OPERATOR:
-                brace_at_caret = caret_pos
-
-        if brace_at_caret >= 0:
-            brace_opposite = self.BraceMatch(brace_at_caret)
-
-        if brace_at_caret != -1 and brace_opposite == -1:
-            self.BraceBadLight(brace_at_caret)
-        else:
-            self.BraceHighlight(brace_at_caret, brace_opposite)
-
-    def on_margin_click(self, evt):
-        # fold and unfold as needed
-        if evt.GetMargin() == 2:
-            if evt.GetShift() and evt.GetControl():
-                self.FoldAll()
-            else:
-                line_clicked = self.LineFromPosition(evt.GetPosition())
-
-                if self.GetFoldLevel(line_clicked) & stc.STC_FOLDLEVELHEADERFLAG:
-                    if evt.GetShift():
-                        self.SetFoldExpanded(line_clicked, True)
-                        self.Expand(line_clicked, True, True, 1)
-                    elif evt.GetControl():
-                        if self.GetFoldExpanded(line_clicked):
-                            self.SetFoldExpanded(line_clicked, False)
-                            self.Expand(line_clicked, False, True, 0)
-                        else:
-                            self.SetFoldExpanded(line_clicked, True)
-                            self.Expand(line_clicked, True, True, 100)
-                    else:
-                        self.ToggleFold(line_clicked)
-
-    def FoldAll(self, action=None):
-        line_count = self.GetLineCount()
-        expanding = True
-
-        # find out if we are folding or unfolding
-        for line_num in range(line_count):
-            if self.GetFoldLevel(line_num) & stc.STC_FOLDLEVELHEADERFLAG:
-                expanding = not self.GetFoldExpanded(line_num)
-                break
-
-        line_num = 0
-
-        while line_num < line_count:
-            level = self.GetFoldLevel(line_num)
-            if level & stc.STC_FOLDLEVELHEADERFLAG and \
-               (level & stc.STC_FOLDLEVELNUMBERMASK) == stc.STC_FOLDLEVELBASE:
-
-                if expanding:
-                    self.SetFoldExpanded(line_num, True)
-                    line_num = self.Expand(line_num, True)
-                    line_num = line_num - 1
-                else:
-                    last_child = self.GetLastChild(line_num, -1)
-                    self.SetFoldExpanded(line_num, False)
-
-                    if last_child > line_num:
-                        self.HideLines(line_num+1, last_child)
-
-            line_num = line_num + 1
-
-    def Expand(self, line, do_expand, force=False, vis_levels=0, level=-1):
-        last_child = self.GetLastChild(line, level)
-        line = line + 1
-
-        while line <= last_child:
-            if force:
-                if vis_levels > 0:
-                    self.ShowLines(line, line)
-                else:
-                    self.HideLines(line, line)
-            else:
-                if do_expand:
-                    self.ShowLines(line, line)
-
-            if level == -1:
-                level = self.GetFoldLevel(line)
-
-            if level & stc.STC_FOLDLEVELHEADERFLAG:
-                if force:
-                    if vis_levels > 1:
-                        self.SetFoldExpanded(line, True)
-                    else:
-                        self.SetFoldExpanded(line, False)
-
-                    line = self.Expand(line, do_expand, force, vis_levels - 1)
-
-                else:
-                    if do_expand and self.GetFoldExpanded(line):
-                        line = self.Expand(line, True, force, vis_levels - 1)
-                    else:
-                        line = self.Expand(line, False, force, vis_levels - 1)
-            else:
-                line = line + 1
-
-        return line
-
-# ----------------------------------------------------------------------
-
 
 class SourceCodeEditor(PythonSTC):
     def __init__(self, parent, style=wx.BORDER_NONE):
